Improve `process_windows_average_time` (#12)
* WIP on avg_surface

* refactor process

* WIP ; use custom process

* Remove the overwrite option

* clean a bit

* Fix error

* work with multiple data file

* correctly identified smalled windows

* add the posibility to filter trials

* Add history to the file

* code clean up

* [bugfix] check if there is event in the file

* remove var nChanel

* oups...

* Prepares code for custom functions

Refactors the code to allow for application of custom functions to trials.
Introduces a new function `apply_function` for this purpose and modifies
`windows_mean_based_on_event` to return all trials instead of the mean.

* Adds custom averaging functions.

Allows users to select from a variety of averaging
functions, including mean, standard deviation, standard
error, and median, in addition to combinations of
mean with standard deviation or standard error.
The selected function is applied to summarise the trial values.